add requirement for deleting a ticket and minor changes to edit one

+++ events/ticket/ticketDelete.js
@@ -1,13 +1,15 @@
-const { CLIENT_ID, BOT_LOGO, BOT_TEXTFOOTER } = require('../../config.json');
-
 module.exports = {
     name : 'interactionCreate',
     async execute(interaction) {
         if (interaction.isButton()) {
             if (interaction.customId === "closeTicket") {
-                interaction.reply({ content: "Ticket closed.", ephemeral: true });
-                await interaction.channel.delete();
-                await interaction.user.send("Your ticket has been closed.");
+                if (interaction.member.roles.cache.has("MODS_ID")) {
+                    interaction.reply({ content: "Ticket closed.", ephemeral: true });
+                    await interaction.channel.delete();
+                    await interaction.user.send("Your ticket has been closed.");
+                } else {
+                    interaction.reply({ content: "You don't have permission to do that.", ephemeral: true });
+                }
             }
         }
     }
